pyproject.toml
[project]
name = "enterprise-intent-router"
version = "0.1.0"
description = "企业 Agent 系统的生产级意图识别与子系统路由原型"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.135.2,<1.0.0",
  "httpx>=0.28.1,<1.0.0",
  "langchain>=1.1.0,<2.0.0",
  "langchain-openai>=1.1.0,<2.0.0",
  "langgraph>=1.0.4,<2.0.0",
  "pydantic>=2.12.5,<3.0.0",
  "python-dotenv>=1.2.1,<2.0.0",
  "uvicorn>=0.42.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.services"]
